Added one more method of package generation.
This method does not generate fit packages following the definition and strategy of the paper. Instead of filling up a package with all minimum items and then virtual items, fill it with all minimum items and if there is still space left, start filling it with the second smallest item and so on. this will ensure that it is not only the smallest item that is squeezed in but all small items. Doing this can, but not necessarily, decrease the total number of packages. It should, however, never be able to INCREASE the number so its only a potential improvement.